Move ISO download card demo to Messages custom message content examples
Also tweak some styling to better align the example with the design
