+++ packages/module/patternfly-docs/content/extensions/chatbot/examples/Messages/UserMessageWithExtraContent.tsx
@@ -2,7 +2,25 @@ import { Fragment, FunctionComponent } from 'react';
 
 import Message from '@patternfly/chatbot/dist/dynamic/Message';
 import userAvatar from './user_avatar.svg';
-import { Alert, Badge, Button, Card, CardBody, CardFooter, CardTitle } from '@patternfly/react-core';
+import {
+  Alert,
+  Badge,
+  Button,
+  ButtonVariant,
+  Card,
+  CardBody,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+  Content,
+  ContentVariants,
+  Flex,
+  FlexItem,
+  Icon
+} from '@patternfly/react-core';
+import ArrowCircleDownIcon from '@patternfly/react-icons/dist/esm/icons/arrow-circle-down-icon';
+import CheckCircleIcon from '@patternfly/react-icons/dist/esm/icons/check-circle-icon';
+import patternflyAvatar from '../Messages/patternfly_avatar.jpg';
 
 const UserActionEndContent = () => {
   // eslint-disable-next-line no-console
@@ -36,6 +54,48 @@ const BeforeMainContent = () => (
   </div>
 );
 
+const downloadCard = (
+  <Card>
+    <CardHeader>
+      <CardTitle>
+        <Icon size="lg" status="success">
+          <CheckCircleIcon />
+        </Icon>
+        Your discovery ISO is ready
+      </CardTitle>
+    </CardHeader>
+
+    <CardBody>
+      <Flex direction={{ default: 'column' }} spaceItems={{ default: 'spaceItemsLg' }}>
+        <FlexItem>
+          <Content component={ContentVariants.p}>
+            To begin adding hosts to your bare metal cluster, you first need to boot them with the generated Discovery
+            ISO. This allows the installation program to see and manage your hardware.
+          </Content>
+        </FlexItem>
+
+        <Flex direction={{ default: 'column' }} spaceItems={{ default: 'spaceItemsSm' }}>
+          <FlexItem>
+            <Button variant={ButtonVariant.primary} icon={<ArrowCircleDownIcon />} isBlock>
+              Download Discovery ISO
+            </Button>
+          </FlexItem>
+
+          <FlexItem alignSelf={{ default: 'alignSelfCenter' }}>
+            <Content component={ContentVariants.small}>1.2 GB • Expires in 24 hours</Content>
+          </FlexItem>
+        </Flex>
+      </Flex>
+    </CardBody>
+
+    <CardFooter>
+      <Content component={ContentVariants.small}>
+        <strong>Next step:</strong> After downloading, boot your bare metal hosts from this ISO image.
+      </Content>
+    </CardFooter>
+  </Card>
+);
+
 export const UserMessageWithExtraContent: FunctionComponent = () => (
   <>
     <Message
@@ -48,6 +108,15 @@ export const UserMessageWithExtraContent: FunctionComponent = () => (
         beforeMainContent: <BeforeMainContent />,
         afterMainContent: <CardInformationAfterMainContent />,
         endContent: <UserActionEndContent />
+      }}
+    />
+    <Message
+      avatar={patternflyAvatar}
+      name="Bot"
+      role="bot"
+      content="All set! I've finished building the Discovery ISO. The next step is to download it and boot your hosts, which you can do using the summary card I've prepared for you below"
+      extraContent={{
+        endContent: downloadCard
       }}
     />
   </>
